Completed the basis of the payments and Stripe
Sheet opens when user selects item, and payment goes through
Renamed a bunch of files to conform with the MVVM layouts

+++ elsa/Utility/PaymentStore.swift
@@ -14,7 +14,7 @@ class PaymentStore: ObservableObject {
     @Published var paymentResult: PaymentResult?
     @Published var isLoading = false
     
-    func preparePayment(uid: String, amount: Int, currency: String) {
+    func preparePayment(uid: String, amount: Double, currency: String) {
         self.paymentSheet = nil
         self.paymentResult = nil
         self.isLoading = true
